--- src/data/getQuestions/index.ts
@@ -1,2 +1,0 @@
-export { getQuestions } from './getQuestions.request'
-export { getQuestionsExample } from './getQuestions.example'
