+++ src/data/index.ts
@@ -1,4 +1,1 @@
-export * from './getQuestion'
-export * from './getQuestions'
-export * from './getExpiredQuestions'
-export * from './getCurrentQuestions'
+export * from './questions'
