+++ src/data/questions/index.ts
@@ -1,0 +1,1 @@
+export { QUESTIONS_EXAMPLE } from './questions.data'
