+++ src/services/getQuestions/index.ts
@@ -1,0 +1,1 @@
+export { getQuestions } from './getQuestions.request'
