+++ src/services/index.ts
@@ -1,0 +1,4 @@
+export * from './getQuestion'
+export * from './getQuestions'
+export * from './getExpiredQuestions'
+export * from './getCurrentQuestions'
